requirements.txt
Django==5.0.2
djangorestframework==3.14.0
django-cors-headers==4.3.1

numpy==1.26.4
scikit-learn==1.4.1

requests==2.31.0
python-dateutil==2.9.0
django-filter==25.2